+++ apis/vara-eth/test/injected.test.ts
@@ -1,50 +1,36 @@
 import { createPublicClient, createWalletClient, webSocket } from 'viem';
-import type { Chain, Hex, PublicClient, WalletClient, WebSocketTransport } from 'viem';
+import type { Account, Chain, Hex, PublicClient, WalletClient, WebSocketTransport } from 'viem';
 import { privateKeyToAccount } from 'viem/accounts';
 import { execSync } from 'node:child_process';
 
-import {
-  EthereumClient,
-  VaraEthApi,
-  getMirrorClient,
-  getRouterClient,
-  getWrappedVaraClient,
-  WsVaraEthProvider,
-  Injected,
-  IInjectedTransaction,
-} from '../src';
+import { EthereumClient, VaraEthApi, getMirrorClient, WsVaraEthProvider, Injected, IInjectedTransaction } from '../src';
 import { hasProps, waitNBlocks } from './common';
 import { config } from './config';
 
 let api: VaraEthApi;
 let publicClient: PublicClient<WebSocketTransport, Chain, undefined>;
-let walletClient: WalletClient<WebSocketTransport>;
+let walletClient: WalletClient<WebSocketTransport, Chain, Account>;
 let ethereumClient: EthereumClient;
 
-let router: ReturnType<typeof getRouterClient>;
 let mirror: ReturnType<typeof getMirrorClient>;
-let wvara: ReturnType<typeof getWrappedVaraClient>;
 
 let programId: `0x${string}`;
 
 beforeAll(async () => {
   const transport = webSocket(config.wsRpc);
 
-  publicClient = createPublicClient<WebSocketTransport, Chain, undefined>({
+  publicClient = createPublicClient({
     transport,
   }) as PublicClient<WebSocketTransport, Chain, undefined>;
   const account = privateKeyToAccount(config.privateKey);
 
-  walletClient = createWalletClient<WebSocketTransport>({
+  walletClient = createWalletClient({
     account,
     transport,
   });
-  ethereumClient = new EthereumClient<WebSocketTransport>(publicClient, walletClient);
-  router = getRouterClient(config.routerId, ethereumClient);
-  const wvaraAddr = await router.wrappedVara();
-  wvara = getWrappedVaraClient(wvaraAddr, ethereumClient);
+  ethereumClient = new EthereumClient(publicClient, walletClient, config.routerId);
 
-  api = new VaraEthApi(new WsVaraEthProvider(), ethereumClient, config.routerId);
+  api = new VaraEthApi(new WsVaraEthProvider(), ethereumClient);
 });
 
 afterAll(async () => {
@@ -95,19 +81,19 @@ describe('Injected Transactions', () => {
     }, 5 * 60_000);
 
     test('should create a correct hash', () => {
-      const injected = new Injected(api.provider, ethereumClient, TX, router);
+      const injected = new Injected(api.provider, ethereumClient, TX);
       expect(injected.hash).toBe(injectedTxHash);
     });
 
     test('should create a correct message id', () => {
-      const injected = new Injected(api.provider, ethereumClient, TX, router);
+      const injected = new Injected(api.provider, ethereumClient, TX);
       expect(injected.messageId).toBe(injectedMessageId);
     });
 
     test('should create a correct signature', async () => {
       const account = privateKeyToAccount(PRIVATE_KEY);
 
-      const injected = new Injected(api.provider, ethereumClient, TX, router);
+      const injected = new Injected(api.provider, ethereumClient, TX);
 
       const signature = await account.sign({ hash: injected.hash });
 
@@ -117,7 +103,7 @@ describe('Injected Transactions', () => {
 
   describe('setup', () => {
     test('should create program', async () => {
-      const tx = await router.createProgram(config.codeId);
+      const tx = await ethereumClient.router.createProgram(config.codeId);
       await tx.sendAndWaitForReceipt();
 
       programId = await tx.getProgramId();
@@ -125,7 +111,7 @@ describe('Injected Transactions', () => {
 
       expect(programId).toBeDefined();
 
-      mirror = getMirrorClient(programId, ethereumClient);
+      mirror = getMirrorClient(programId, walletClient, publicClient);
     });
 
     test(
@@ -149,7 +135,7 @@ describe('Injected Transactions', () => {
     );
 
     test('should approve wvara', async () => {
-      const tx = await wvara.approve(programId, BigInt(10 * 1e12));
+      const tx = await ethereumClient.wvara.approve(programId, BigInt(10 * 1e12));
 
       await tx.send();
 
@@ -159,7 +145,7 @@ describe('Injected Transactions', () => {
 
       expect(approvalData.value).toEqual(BigInt(10 * 1e12));
 
-      const allowance = await wvara.allowance(ethereumClient.accountAddress, programId);
+      const allowance = await ethereumClient.wvara.allowance(ethereumClient.accountAddress, programId);
       expect(allowance).toEqual(BigInt(10 * 1e12));
     });